Views to list and get Users.
Go to `localhost/users/` to list all Users.
Go to `localhost/users/id` to get a specific User.
If you request HTML from those pages it'll spit out a nice HTML view for
listing and editing Users. If the frontend JS requests JSON it'll spit
out a JSON representation of the Users/User.