Weight chart is now a directive
Just put `<weight-chart></weight-chart>` where you want a chart.
Right now it's just using dummy data, I'll hook it up to the current
user and hit the API next. I'm using NVD3.js, which is nice and handles
a lot of the responsive stuff, but it's not as powerful as just plain
D3. But plain D3 is annoying to make responsive... I might give up on
responsiveness and just do plain D3 stuff so we can have really powerful
and interactive charts.